refactor : Spring Security Refactoring - 불필요한 설정 제거 - 불필요한 리소스 사용 제거 - TokenAuthenticationFilter, TokenProvider 현재 미사용 - OAuth2SuccessHandler 토큰 관련 로직 주석 처리
